Remove Python version check warning
